feat: cart shows total price
